Moved the generic parsing code to `mod parse`.
