+++ analysis/runtime/src/lib.rs
@@ -3,6 +3,7 @@ mod handlers;
 pub mod metadata;
 pub mod mir_loc;
 pub mod runtime;
+mod parse;
 
 pub use handlers::*;
 use runtime::{global_runtime::RUNTIME, skip::notify_if_events_were_skipped_before_main};
